Initial commit
Add these files to the same project you want to check the resx file. Update the values of aspxContent and aspxCsContent in WebForm1.aspx.cs. Build, and open WebForm1.aspx in a browser to get a list of the items that are referenced on on the aspx and cs files. Note; not all use cases have been determined, but it's a start.